api/requirements.txt
flask==2.2.2
flask-cors==3.0.10
python-dotenv==0.20.0
langchain_core==0.1.17
langchain_google_genai==0.1.0
